refactor: split render.rs monolith into focused modules
Extracts the ~950 line src/render.rs into a src/render/ directory with:
- table.rs for comfy-table rendering
- csv.rs for CSV export
- json.rs for JSON export
- helpers.rs for colorization and formatting
- mod.rs to maintain the public API
